update album by id

+++ models/album.go
@@ -1,5 +1,6 @@
 package models
 
+// Album represents album data stored in the database
 type Album struct {
 	ID     int64
 	Title  string
@@ -7,9 +8,10 @@ type Album struct {
 	Price  float32
 }
 
+// AlbumDTO is the data transfer object for Album
 type AlbumDTO struct {
 	ID     string  `json:"id"`
 	Title  string  `json:"title"`
 	Artist string  `json:"artist"`
-	Price  float32 `json:"price"`
+	Price  float64 `json:"price"`
 }
